Renamed _robust_decomposition back to robust_decomposition
