+++ tensorly/decomposition/__init__.py
@@ -6,7 +6,7 @@ tensor decomposition such as CANDECOMP-PARAFAC and Tucker.
 from ._cp import (parafac, CP, RandomizedCP, randomised_parafac, sample_khatri_rao)
 from ._nn_cp import non_negative_parafac
 from ._tucker import tucker, partial_tucker, non_negative_tucker, Tucker
-from ._robust_decomposition import robust_pca
+from .robust_decomposition import robust_pca
 from ._tt import TensorTrain, tensor_train, tensor_train_matrix
 from ._parafac2 import parafac2, Parafac2
 from ._symmetric_cp import symmetric_parafac_power_iteration, symmetric_power_iteration, SymmetricCP
